[JS importmap] Configure importmap for the rider_app engine
The configuration showcased here is meant to mimic 100% the structure and feel of a vanilla rails app.
For example, when a controller is declared in html it's declared without any prefixes.

I believe the folder configuration "javascript/rider_app" is necessary to distinguish importmap setups between engines.

It's very important to properly configure the importmap file to load stimulus controllers correctly. This is a two step process:
1. Tell the rails asset manager (Propshaft in this case) to compile assets for the engine using an initializer in the engine.rb file
2. Make sure to add the proper arguments to `pin_all_from`
  - path: use `File.expand_path(..., __dir__)` with the path to the controllers folder
  - under: THIS IS VERY IMPORTANT and took me a long while to figure out. Basically, this argument prefixes the file name to make sure it's correctly found. For example, if the file name is `hello_controller.js` the "under" argument will prefix it so it's correctly found after assets are compiled: "rider_app/controllers/hello_controller.js". This works in conjunction with the point 1 above.