[platform package] Extract ApplicationRecord in platform engine
ApplicationRecord is usually the `primary_abstract_class` for most models to use in a single-db setup. By extracting it to a platform-level package we encourage all children to use this class. Contrary to the initial believe that this shouldn't be a proper engine because it doesn't quick mimic an app, I found that not making it an engine means it doesn't have the benefit of autoloading classes when the rails app is initialized and creates problems with packwerk. You may be able to circumvnet this but autoloading constants manually but that is a bit of a contradiction: manually autoloading something.

I didn't run into this issue in other apps at work because packages were defined inside of the Rails app `app` folder which is autoloaded but because this structure relies on having stuff outisde of `app` I found this issue.