[multi-engine example] Create Driver app
This PR creates the Driver App as a multi-engine app that compartmentalizes features using engines. The way this app is organized is special in that it showcases the power of having custom folder structures that are not picked up by the autoloader, allowing convenient folder nesting without creating unnecessary ruby modules. To exemplify this, look at the structure of the "features" folder inside of "driver_app". In any regular rails app, if an features folder was put inside of "app", it would create a namespace, but if it wasn't, it wouldn't be picked up by the autoloader. If we created added something like elefante/features to the autoloader, we'd then have to tell it to ignore the first level of folders to have a similar structure to a rails app with models/jobs etc.

With this setup, folder structure and autoloading are separated and the latter is left to rails' default rules for engines.