[DX] Enhance readability with custom editor configurations
This architecture comes with a DX setback: complex folder nesting.
I realised though that good architecture can't be overshadowed by a poor editor experience. That's why this PR includes custom workspaces/scopes for Rubymine an VSCode to make the developer experience better.
The objective is to provide zoomed-in views into the packages/engines the developer's working on and hiding other packages/engines.