[packwerk] Add package.yml to all packages
This ensures all dependent code is compliant. Before this commit, running `bin/packwerk check` would shocase 4 errors:
2 privacy ones and 2 dependency ones.

Privacy:
Initially, the application_record.rb file was defined inside `app/models` as it traditionally is. This means that, in the eyes of packwerk, this is a private, package-level file which should not be used by the outside world. By putting this file inside the `app/public` folder (packwerk's default folder for public interfaces) we remove the 2 privacy violations.

The 2 dependency violations are fixed by making sure both the driver and rider apps declare that they depend on the platform package.

In the future, the platform package could have other "ApplicationRecords" that target other databases for a multi-db setup. Maybe this package could be renamed to "persistence" where all "record" parents are defined.